Move services related to reconcile in separate package
