+++ axelor-account/src/main/java/com/axelor/apps/account/db/repo/ReconcileManagementRepository.java
@@ -21,7 +21,7 @@ package com.axelor.apps.account.db.repo;
 import com.axelor.apps.account.db.MoveLine;
 import com.axelor.apps.account.db.Reconcile;
 import com.axelor.apps.account.exception.AccountExceptionMessage;
-import com.axelor.apps.account.service.ReconcileSequenceService;
+import com.axelor.apps.account.service.reconcile.ReconcileSequenceService;
 import com.axelor.apps.base.AxelorException;
 import com.axelor.apps.base.db.repo.TraceBackRepository;
 import com.axelor.apps.base.service.exception.TraceBackService;
