+++ axelor-account/src/main/java/com/axelor/apps/account/module/AccountModule.java
@@ -330,6 +330,20 @@ import com.axelor.apps.account.service.payment.paymentvoucher.PayVoucherDueEleme
 import com.axelor.apps.account.service.payment.paymentvoucher.PayVoucherDueElementServiceImpl;
 import com.axelor.apps.account.service.payment.paymentvoucher.PaymentVoucherCancelService;
 import com.axelor.apps.account.service.payment.paymentvoucher.PaymentVoucherCancelServiceImpl;
+import com.axelor.apps.account.service.reconcile.ReconcileService;
+import com.axelor.apps.account.service.reconcile.ReconcileServiceImpl;
+import com.axelor.apps.account.service.reconcile.reconcilegroup.ReconcileGroupService;
+import com.axelor.apps.account.service.reconcile.reconcilegroup.ReconcileGroupServiceImpl;
+import com.axelor.apps.account.service.reconcilegroup.ReconcileGroupFetchService;
+import com.axelor.apps.account.service.reconcilegroup.ReconcileGroupFetchServiceImpl;
+import com.axelor.apps.account.service.reconcilegroup.ReconcileGroupLetterService;
+import com.axelor.apps.account.service.reconcilegroup.ReconcileGroupLetterServiceImpl;
+import com.axelor.apps.account.service.reconcilegroup.ReconcileGroupProposalService;
+import com.axelor.apps.account.service.reconcilegroup.ReconcileGroupProposalServiceImpl;
+import com.axelor.apps.account.service.reconcilegroup.ReconcileGroupSequenceService;
+import com.axelor.apps.account.service.reconcilegroup.ReconcileGroupSequenceServiceImpl;
+import com.axelor.apps.account.service.reconcilegroup.ReconcileGroupUnletterService;
+import com.axelor.apps.account.service.reconcilegroup.ReconcileGroupUnletterServiceImpl;
 import com.axelor.apps.account.service.umr.UmrNumberService;
 import com.axelor.apps.account.service.umr.UmrNumberServiceImpl;
 import com.axelor.apps.account.util.TaxAccountToolService;
