+++ axelor-account/src/main/java/com/axelor/apps/account/service/reconcile/reconcilegroup/ReconcileGroupServiceImpl.java
@@ -16,7 +16,7 @@
  * You should have received a copy of the GNU Affero General Public License
  * along with this program.  If not, see <https://www.gnu.org/licenses/>.
  */
-package com.axelor.apps.account.service;
+package com.axelor.apps.account.service.reconcile.reconcilegroup;
 
 import com.axelor.apps.account.db.Account;
 import com.axelor.apps.account.db.MoveLine;
@@ -26,6 +26,8 @@ import com.axelor.apps.account.db.repo.MoveLineRepository;
 import com.axelor.apps.account.db.repo.ReconcileGroupRepository;
 import com.axelor.apps.account.db.repo.ReconcileRepository;
 import com.axelor.apps.account.exception.AccountExceptionMessage;
+import com.axelor.apps.account.service.reconcilegroup.ReconcileGroupFetchService;
+import com.axelor.apps.account.service.reconcilegroup.ReconcileGroupSequenceService;
 import com.axelor.apps.base.AxelorException;
 import com.axelor.apps.base.db.Company;
 import com.axelor.apps.base.db.repo.TraceBackRepository;
