+++ axelor-account/src/main/java/com/axelor/apps/account/service/reconcilegroup/ReconcileGroupLetterService.java
@@ -1,4 +1,4 @@
-package com.axelor.apps.account.service;
+package com.axelor.apps.account.service.reconcilegroup;
 
 import com.axelor.apps.account.db.ReconcileGroup;
 import com.axelor.apps.base.AxelorException;
