+++ axelor-account/src/main/java/com/axelor/apps/account/service/reconcilegroup/ReconcileGroupUnletterServiceImpl.java
@@ -1,7 +1,8 @@
-package com.axelor.apps.account.service;
+package com.axelor.apps.account.service.reconcilegroup;
 
 import com.axelor.apps.account.db.Reconcile;
 import com.axelor.apps.account.db.ReconcileGroup;
+import com.axelor.apps.account.service.reconcile.ReconcileService;
 import com.axelor.apps.base.AxelorException;
 import com.google.inject.Inject;
 import com.google.inject.persist.Transactional;
