+++ axelor-account/src/main/java/com/axelor/apps/account/web/BatchAccountController.java
@@ -24,8 +24,8 @@ import com.axelor.apps.account.db.ReconcileGroup;
 import com.axelor.apps.account.db.repo.MoveLineRepository;
 import com.axelor.apps.account.db.repo.ReconcileGroupRepository;
 import com.axelor.apps.account.exception.AccountExceptionMessage;
-import com.axelor.apps.account.service.ReconcileGroupProposalService;
 import com.axelor.apps.account.service.batch.BatchControlMovesConsistency;
+import com.axelor.apps.account.service.reconcilegroup.ReconcileGroupProposalService;
 import com.axelor.apps.base.AxelorException;
 import com.axelor.apps.base.ResponseMessageType;
 import com.axelor.apps.base.db.Batch;
