+++ axelor-account/src/main/java/com/axelor/apps/account/web/ReconcileController.java
@@ -20,7 +20,7 @@ package com.axelor.apps.account.web;
 
 import com.axelor.apps.account.db.Reconcile;
 import com.axelor.apps.account.db.repo.ReconcileRepository;
-import com.axelor.apps.account.service.ReconcileService;
+import com.axelor.apps.account.service.reconcile.ReconcileService;
 import com.axelor.apps.base.ResponseMessageType;
 import com.axelor.apps.base.service.exception.TraceBackService;
 import com.axelor.inject.Beans;
