+++ axelor-account/src/main/java/com/axelor/apps/account/web/ReconcileGroupController.java
@@ -21,9 +21,9 @@ package com.axelor.apps.account.web;
 import com.axelor.apps.account.db.MoveLine;
 import com.axelor.apps.account.db.ReconcileGroup;
 import com.axelor.apps.account.db.repo.ReconcileGroupRepository;
-import com.axelor.apps.account.service.ReconcileGroupLetterService;
-import com.axelor.apps.account.service.ReconcileGroupProposalService;
-import com.axelor.apps.account.service.ReconcileGroupUnletterService;
+import com.axelor.apps.account.service.reconcilegroup.ReconcileGroupLetterService;
+import com.axelor.apps.account.service.reconcilegroup.ReconcileGroupProposalService;
+import com.axelor.apps.account.service.reconcilegroup.ReconcileGroupUnletterService;
 import com.axelor.apps.base.AxelorException;
 import com.axelor.apps.base.ResponseMessageType;
 import com.axelor.apps.base.service.exception.ErrorException;
