+++ axelor-account/src/main/java/com/axelor/csv/script/ImportReconcile.java
@@ -20,7 +20,7 @@ package com.axelor.csv.script;
 
 import com.axelor.apps.account.db.Reconcile;
 import com.axelor.apps.account.db.repo.ReconcileRepository;
-import com.axelor.apps.account.service.ReconcileServiceImpl;
+import com.axelor.apps.account.service.reconcile.ReconcileServiceImpl;
 import com.axelor.apps.base.AxelorException;
 import com.google.inject.Inject;
 import java.util.Map;
